cpu-boost: Implement Two-Step input boost
The expected cpu demand is higher right after an
input event. The longer ago the last input away is,
the lower is the expected cpu demand.

Based on that a assumption, it makes sense to implement
a two-step input boost, where both steps are fully
configurable.

This is a minimal implementation of a two-step input
boost.

It is disabled by default and gets enabled automatically
by configuring the parameters input_boost_freq_s2 and
input_boost_ms_s2. These parameters represent step 2.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Ahmed Radaideh <ahmed.radaideh@gmail.com>