BACKPORT: genirq: Allow migration and naming of chained IRQs
This change addresses the lack of a name for IRQ 354 on
marlin/sailfish devices in both kernel and batterystats logs.  It is
comprised of three backported changes: two external kernel changes
that modified registration of chained IRQ handlers to all use a shared
generic irqaction, and an internal change that built upon that to
provide a name for that shared irqaction so chained IRQs will be
logged properly.  Because of the nature of the change, the name
provided ("chained-irq") has to be quite generic.

(The primary purpose of the external changes was to allow migration of
chained handlers away from offlined CPUs, which required that those IRQs
have non-null actions.)

Backported changes are:
commit e509bd7da149 ("genirq: Allow migration of chained interrupts by
installing default action")
commit 4717f133736d ("genirq/PM: Restore system wake up from chained
interrupts")
commit c14d789defb0 ("genirq: Add generic action name for chained IRQs")

Bug: 63899519
Fixes: 34961013
Test: Compared before/after kernel logs and batterystats output on both
marlin and sailfish devices to verify efficacy.
Test: Ran on marlin and sailfish devices for multiple days to verify
stability with chained IRQ migration.
Signed-off-by: Kelly Rossmoyer <krossmo@google.com>

Change-Id: Idaddeaa0ef02d5130a1f60e7b472338279a97ec9
Signed-off-by: Ahmed Radaideh <ahmed.radaideh@gmail.com>