mm: swap: swap pages one at a time
According to Google, this is optimal.
"By default, the Linux kernel swaps in 8 pages of memory at a time. When
using ZRAM, the incremental cost of reading 1 page at a time is
negligible and may help in case the device is under extreme memory
pressure."

Source: source.android.com/devices/tech/perf/low-ram